backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
requests==2.32.3
python-multipart==0.0.19
pillow==11.0.0
pytest==8.3.4
httpx==0.28.1
